fix(admin): cast avg fulfillment hours to float8 in stats
EXTRACT(EPOCH FROM AVG(...)) returns numeric on PG14+, which the pg driver serializes as a JSON string. /api/admin/stats and the web stats page passed avgFulfillmentHours through uncast, so a typed client (the iOS app) failed to decode it as a number and the whole Statistics screen errored with a decode failure. Cast to ::float8 like every other float raw-query in the repo so the value stays JSON-numeric.

+++ src/app/(app)/admin/stats/page.tsx
@@ -29,7 +29,7 @@ export default async function StatsPage() {
     prisma.jellyfinLibraryItem.groupBy({ by: ["mediaType"], _count: { _all: true } }),
     prisma.tVEpisodeCache.groupBy({ by: ["source"], _count: { _all: true }, _sum: { runtime: true } }),
     prisma.$queryRaw<{ avg_hours: number | null }[]>`
-      SELECT EXTRACT(EPOCH FROM AVG("availableAt" - "createdAt")) / 3600 AS avg_hours
+      SELECT (EXTRACT(EPOCH FROM AVG("availableAt" - "createdAt")) / 3600)::float8 AS avg_hours
       FROM "MediaRequest"
       WHERE status = 'AVAILABLE' AND "availableAt" IS NOT NULL
     `,
